WELD-2796 Bundle imports for built-in bean proxies
When a client proxy is created for a built-in bean, it may be loaded
using the classloader of the weld osgi bundle because it provides the
built-in bean.

This means that the bundle should import all the packages necessary to
load the proxy class, which may include packages that the code in the
bundle doesn't directly reference.

The following package imports are added:
- jakarta.servlet.descriptor (for ServletContext)
- javax.security.auth (for Principal)